+++ NEWS.md
@@ -3,7 +3,8 @@
 * Some improvements to form "Burn Grading 0.1" identified and in preparation
   as "Burn Grading 0.2". Pending review and QA.
 * Add workhorse function `calculate_geocbi`. Example see vignette "analysis".
-* Move helper `map_burngrading` from vignette to own function.
+* Move helpers `map_burngrading` and `reactable_burngrading` from vignette 
+  to own functions.
 * Add Valerie Densmore as package author. Valerie has refined the GeoCBI for
   Western Austrlian Flora.